Add sass-loader
