+++ package.json
@@ -18,8 +18,13 @@
   },
   "homepage": "https://github.com/kubilaycaglayan/Restaurant-Page#readme",
   "devDependencies": {
+    "css-loader": "^3.6.0",
     "file-loader": "^6.0.0",
+    "sass": "^1.26.9",
+    "sass-loader": "^8.0.2",
+    "style-loader": "^1.2.1",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.12"
-  }
+  },
+  "dependencies": {}
 }
