+++ src/compHeader.js
@@ -1,8 +1,11 @@
+import navbar from './compNav';
+
 export default function header() {
   const header = document.createElement('header');
   const h1 = document.createElement('h1');
   h1.innerHTML = 'Welcome To The Michelin\'s Place';
   h1.className = 'display-4';
+  header.appendChild(navbar());
   header.appendChild(h1);
   header.className = 'bg-danger py-5 text-center text-white';
   return header;
