+++ src/compMain.js
@@ -1,26 +1,6 @@
 import MichelinStar from './michelin.png';
 import Dish from './dish1.png';
-
-function createElement(element, className) {
-  const elm = document.createElement(element);
-  elm.className = className;
-  return elm;
-}
-
-function createElementWithText(element, className, text) {
-  const elm = document.createElement(element);
-  elm.className = className;
-  elm.innerHTML = text;
-  return elm;
-}
-
-function createImage(src, alt, className) {
-  const myImage = new Image();
-  myImage.src = src;
-  myImage.alt = alt;
-  myImage.className = className;
-  return myImage;
-}
+import { createElement, createElementWithText, createImage } from './functions';
 
 export default function main() {
   const childrenOfMain = [];
