+++ src/compNav.js
@@ -1,0 +1,7 @@
+import { createElement, createElementWithText, createImage } from './functions';
+
+export default function navbar() {
+  const navbar = createElement('navbar', 'try');
+  navbar.innerHTML = 'Merhaba';
+  return navbar;
+}
