+++ src/style.sass
@@ -1,0 +1,2 @@
+.try
+  color: red
